AI_Engine/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pandas==2.2.2
numpy==1.26.4
xgboost==2.1.0
scikit-learn==1.5.0
joblib==1.4.2
rich
python-dotenv
